pyproject.toml
virtualenvs.in-project = true

pythonpath="witch"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.poetry]
name = "witch"
version = "0.1.0"
description = ""
authors = ["Rob Spaul"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.13"
#fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlmodel = "^0.0.22"
python-fasthtml = "^0.12.0"
#flask = "^3.1.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", # requires docstring for public modules
    "D104", # requires docstring for public packages
    "D401", # imperative mood
    "D211",
    "D213",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = [
    "S101",  # use of assert (required for pytest)
    "D103",  # requiring docstring for unit tests
]

[tool.ruff.format]
quote-style = "double"  # Like Black, use double quotes for strings.
indent-style = "space"  # Like Black, indent with spaces, rather than tabs.
skip-magic-trailing-comma = false  # Like Black, respect magic trailing commas.
line-ending = "auto"  # Like Black, automatically detect the appropriate line ending.
docstring-code-format = true
